requirements.txt
amqp==2.5.1
atomicwrites==1.3.0
attrs==19.1.0
python-dotenv==0.10.3
backcall==0.1.0
billiard==3.5.0.2
bleach==3.1.0
blessings==1.7
cameo==0.10.3
celery==4.0.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cobra==0.9.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
depinfo==1.5.1
dill==0.2.5
docutils==0.15.2
entrypoints==0.3
Escher==1.7.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Cors==3.0.8
Flask-JWT==0.3.2
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.2
flask-swagger==0.2.14
Flask-Testing==0.7.1
Flask-RESTful==0.3.5
future==0.17.1
gnomic==1.0.1
grako==3.18.1
idna==2.8
importlib-metadata==0.19
inspyred==1.0.1
IProgress==0.4
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
kombu==4.0.2
lazy-object-proxy==1.4.2
lockfile==0.12.2
MarkupSafe==1.1.1
marshmallow==2.13.0
marshmallow-sqlalchemy==0.12.1
matplotlib==3.0.3
metabolitics==0.0.12
mistune==0.8.4
more-itertools==7.2.0
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==1.9.1
nltk==3.2.4
notebook==6.0.1
numexpr==2.7.0
numpy==1.16.4
openpyxl==2.6.3
optlang==1.4.4
ordered-set==3.1.1
packaging==19.1
palettable==3.2.0
pandas==0.21.1
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.4
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pid==2.2.5
pipdeptree==0.13.2
plotly==3.10.0
pluggy==0.12.0
preprocessing==0.1.13
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.1
ptyprocess==0.6.0
py==1.8.0
PyFunctional==1.2.0
Pygments==2.4.2
PyJWT==1.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==4.6.5
python-daemon==2.2.3
python-dateutil==2.8.0
python-libsbml-experimental==5.18.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.1.0
redis==2.10.5
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.14.12
ruamel.yaml.clib==0.1.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
flower==0.9.2
service==0.6.0
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
gunicorn==19.7.0
SQLAlchemy-JSONField==0.6.4
sklearn-utils==0.0.15
sphinx-rtd-theme==0.2.4
SQLAlchemy==1.1.6
statsmodels==0.10.1
swiglpk==4.65.0
sympy==1.4
tabulate==0.8.3
terminado==0.8.2
babel == 2.3.4
testpath==0.4.2
tornado==4.5.1
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.0
widgetsnbextension==3.5.1
WTForms==2.2.1
zipp==0.6.0
